forum design

+++ JoinUs/src/main/java/com/northgatecode/joinus/mongodb/Category.java
@@ -10,22 +10,22 @@ import org.mongodb.morphia.annotations.Id;
 @Entity(noClassnameStored = true)
 public class Category {
     @Id
-    private long id;
+    private int id;
     private String name;
 
     public Category() {
     }
 
-    public Category(long id, String name) {
+    public Category(int id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
